Initial Code Revision
Added index and packages.
Added express package reference
Index.js sets up a default route for "Hello World"-ing and sets up the
necessary routes to respond to. Right now the response for each route
simply returns the parameters it was passed. This is a decent start to
the service. Next steps are to write the solving module and wire the web
service to call it.